Tests: Moved ServiceProvider::setExpects to TestCase
